+++ CHANGELOG.md
@@ -47,6 +47,9 @@ Bugsnag Notifiers on other platforms.
   `BugsnagConfiguration` value but can be overridden in event passed to the 
   `Bugsnag.notify()` callback.
   [#458](https://github.com/bugsnag/bugsnag-cocoa/pull/458)
+  
+* Add metadata accessor methods to `BugsnagEvent`
+  []()
 
 ## Bug fixes